Added src folder with code for F1 world Champion
I have used http://ergast.com/mrd/ to create a single page application that presents a list that shows the F1 world champions starting from 2005 until 2015. Clicking on an item shows the list of the winners for every race for the selected year. I have also highlighted the row when the winner has been the world champion in the same season.